Spawn FlockManager for new pngs in InputImages
Rather than starting with a flock of boids in the scene, we instead
spawn the FlockManager when a new .png file is added to the
InputImages folder. We use the png as the sprite for the boids.

The idea is that we'll use TouchDesigner to generate visuals,
and dump them into that folder, and use Unity to handle
the flocking boids.

This may be extended to use animations, rather than still sprites,
for the boids. But this is just a starting point.